feat(metastore): metastore DI
This commit fixes a problem that happens when we instantiate 2 metastores in a single process.
Even though this can't happen at the moment it will be the case once we introduce distributed
metastore queries - when running both schedulers and workers in the same process all of them
will need an instance of metastore. Each instance (before this PR) internally creates and
registers metastore metrics which causes panic due to exactly the same metrics were already
registered by the same registerer.

I attempt to fix this problem via creating and registering metastore metrics only once.